+++ README.md
@@ -29,7 +29,7 @@
 - ✅ 解决 Wayland 环境下渲染黑屏问题
 - ✅ 支持手动选择路径与语音包管理
 
-> **注意**：Linux 用户请务必查看 [Linux 使用指南](LINUX.md) 以安装必要依赖和配置环境变量。
+> **注意**：Linux 用户请务必查看 [Linux 使用指南](docs/LINUX.md) 以安装必要依赖和配置环境变量。
 
 ## 快速开始（源码运行）
