+++ docs/info_template.txt
