+++ services/config_manager.py
@@ -12,7 +12,7 @@ import os
 import platform
 from pathlib import Path
 import sys
-from logger import get_logger
+from utils.logger import get_logger
 
 log = get_logger(__name__)
