+++ services/manifest_manager.py
@@ -14,7 +14,7 @@ import json
 from pathlib import Path
 from datetime import datetime
 from typing import Any
-from logger import get_logger
+from utils.logger import get_logger
 
 log = get_logger(__name__)
