+++ services/sights_manager.py
@@ -29,7 +29,7 @@ import subprocess
 import zipfile
 from pathlib import Path
 from typing import Callable, Any
-from logger import get_logger
+from utils.logger import get_logger
 
 log = get_logger(__name__)
