+++ services/skins_manager.py
@@ -25,7 +25,7 @@ import zipfile
 from pathlib import Path
 from typing import Callable, Any
 
-from logger import get_logger
+from utils.logger import get_logger
 
 log = get_logger(__name__)
