+++ services/telemetry_manager.py
